Move files for default Python folder structure
