+++ test/test_device.py
@@ -24,7 +24,7 @@ import warnings
 
 from pathlib import Path
 
-from elitech.src.device import Device
+from elitech.device import Device
 
 
 class TestDevice(unittest.TestCase):
@@ -68,7 +68,7 @@ class TestDevice(unittest.TestCase):
             dev = Device(Path('/dev/non-existing-device'))
         self.assertEqual(str(e.exception), "Device \"/dev/non-existing-device\" does not exist")
 
-    @unittest.mock.patch('elitech.src.device.open')
+    @unittest.mock.patch('elitech.device.open')
     def testOutReportSize(self, mock_open):
         with open(Path(__file__).parents[0] / 'data' / 'hid_report_descriptor', 'rb') as f:
             unittest.mock.mock_open(mock_open, f.read())
@@ -87,7 +87,7 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[0][0][0], Path('/sys/class/hidraw/null/device/report_descriptor'))
         self.assertEqual(mock_open.call_args_list[0][0][1], 'rb')
 
-    @unittest.mock.patch('elitech.src.device.open')
+    @unittest.mock.patch('elitech.device.open')
     def testInReportSize(self, mock_open):
         with open(Path(__file__).parents[0] / 'data' / 'hid_report_descriptor', 'rb') as f:
             unittest.mock.mock_open(mock_open, f.read())
@@ -106,7 +106,7 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[0][0][0], Path('/sys/class/hidraw/null/device/report_descriptor'))
         self.assertEqual(mock_open.call_args_list[0][0][1], 'rb')
 
-    @unittest.mock.patch('elitech.src.device.print')
+    @unittest.mock.patch('elitech.device.print')
     def testWriteEmpty(self, mock_print):
         dev = Device('')
         dev.write(bytes([b for b in range(0, 11)]))
@@ -114,7 +114,7 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(len(mock_print.call_args_list), 1)
         self.assertEqual(mock_print.call_args_list[0][0][0], 'Request:  ' + ' '.join([(f'{b:02X}' if b < 11 else '00') for b in range(0, 64)]))
 
-    @unittest.mock.patch('elitech.src.device.print')
+    @unittest.mock.patch('elitech.device.print')
     def testReadEmpty(self, mock_print):
         dev = Device('')
         self.assertEqual(dev.read(), bytes([0]*64))
@@ -122,7 +122,7 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(len(mock_print.call_args_list), 1)
         self.assertEqual(mock_print.call_args_list[0][0][0], 'Response: ' + ' '.join(['00']*64))
 
-    @unittest.mock.patch('elitech.src.device.open')
+    @unittest.mock.patch('elitech.device.open')
     def testOpen(self, mock_open):
         mock_file = unittest.mock.Mock()
         mock_open.return_value = mock_file
@@ -138,8 +138,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[0][0][1], 'rb+')
         self.assertEqual(len(mock_file.close.call_args_list), 1)
 
-    @unittest.mock.patch('elitech.src.device.print')
-    @unittest.mock.patch('elitech.src.device.open')
+    @unittest.mock.patch('elitech.device.print')
+    @unittest.mock.patch('elitech.device.open')
     def testWrite(self, mock_open, mock_print):
         mock_descriptor = unittest.mock.MagicMock()
         mock_descriptor.__enter__.return_value = mock_descriptor
@@ -167,8 +167,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(len(mock_print.call_args_list), 1)
         self.assertEqual(mock_print.call_args_list[0][0][0], 'Request:  ' + ' '.join([(f'{b:02X}' if b < 11 else '00') for b in range(0, 64)]))
 
-    @unittest.mock.patch('elitech.src.device.print')
-    @unittest.mock.patch('elitech.src.device.open')
+    @unittest.mock.patch('elitech.device.print')
+    @unittest.mock.patch('elitech.device.open')
     def testRead(self, mock_open, mock_print):
         mock_descriptor = unittest.mock.MagicMock()
         mock_descriptor.__enter__.return_value = mock_descriptor
@@ -197,8 +197,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(len(mock_print.call_args_list), 1)
         self.assertEqual(mock_print.call_args_list[0][0][0], 'Response: ' + ' '.join([f'{b:02X}' for b in range(0, 64)]))
 
-    @unittest.mock.patch('elitech.src.device.print')
-    @unittest.mock.patch('elitech.src.device.open')
+    @unittest.mock.patch('elitech.device.print')
+    @unittest.mock.patch('elitech.device.open')
     def testReadKeyboardInterrupt(self, mock_open, mock_print):
         mock_descriptor = unittest.mock.MagicMock()
         mock_descriptor.__enter__.return_value = mock_descriptor
@@ -227,8 +227,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(len(mock_print.call_args_list), 1)
         self.assertEqual(mock_print.call_args_list[0][0][0], 'Response: ' + ' '.join(['00' for b in range(0, 64)]))
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {'null': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/null'}},
             'devices': {'pci0000:00': {'0000:00:14.0': {'usb3': {'3-2': {
@@ -249,8 +249,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {
                 'null': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/null'},
@@ -274,8 +274,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {
                 'hidraw0': b'',
@@ -299,8 +299,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {'null': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/null'}},
             'devices': {'pci0000:00': {'0000:00:14.0': {'usb3': {'3-2': {
@@ -319,8 +319,8 @@ class TestDevice(unittest.TestCase):
             dev.productId
         self.assertEqual(str(e.exception), f"Device vendor id and product id cannot be obtained")
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {
                 'hidraw0': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/hidraw0',
@@ -344,8 +344,8 @@ class TestDevice(unittest.TestCase):
             dev.productId
         self.assertEqual(str(e.exception), f"Device vendor id and product id cannot be obtained")
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {'null': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/null'}},
             'devices': {'pci0000:00': {'0000:00:14.0': {'usb3': {'3-2': {
@@ -365,8 +365,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {
                 'null': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/null'},
@@ -389,8 +389,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {
                 'hidraw0': b'',
@@ -414,8 +414,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {'null': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/null'}},
             'devices': {'pci0000:00': {'0000:00:14.0': {'usb3': {'3-2': {
@@ -438,8 +438,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {
                 'null': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/null'},
@@ -465,8 +465,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'sys' : {
             'class': {'hidraw': {
                 'hidraw0': b'',
@@ -492,8 +492,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(mock_open.call_args_list[1][0][0], Path('/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/idProduct'))
         self.assertEqual(mock_open.call_args_list[1][0][1], 'rt')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'dev': {'hidraw0': b''},
         'sys': {
             'class': {'hidraw': {'hidraw0': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/hidraw0'}},
@@ -505,8 +505,8 @@ class TestDevice(unittest.TestCase):
     def testEnumerateError(self, mock_path, mock_open):
         self.assertEqual(len([d for d in Device.enumerate()]), 0)
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'dev': {'hidraw0': b''},
         'sys': {
             'class': {'hidraw': {'hidraw0': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/hidraw0'}},
@@ -520,8 +520,8 @@ class TestDevice(unittest.TestCase):
     def testEnumerateNone(self, mock_path, mock_open):
         self.assertEqual(len([d for d in Device.enumerate()]), 0)
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'dev': {'hidraw0': b''},
         'sys': {
             'class': {'hidraw': {'hidraw0': '/sys/devices/pci0000:00/0000:00:14.0/usb3/3-2/3-2:1.0/0003:046D:C069.0002/hidraw/hidraw0'}},
@@ -540,8 +540,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(devices[0].productId, 0x3005)
         self.assertEqual(devices[0].name, 'Elitech RC-5+')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'dev': {
             'hidraw0': b'',
             'hidraw1': b'',
@@ -573,8 +573,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(devices[0].productId, 0x3005)
         self.assertEqual(devices[0].name, 'Elitech RC-5+')
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'dev': {
             'hidraw0': b'',
             'hidraw1': b'',
@@ -607,8 +607,8 @@ class TestDevice(unittest.TestCase):
         self.assertEqual(devices[0].name, 'Elitech RC-5+')
 
 
-    @unittest.mock.patch('elitech.src.device.open', new_callable=mockpath.MockPath.mock_open)
-    @unittest.mock.patch('elitech.src.device.Path', new_callable=mockpath.MockPath({
+    @unittest.mock.patch('elitech.device.open', new_callable=mockpath.MockPath.mock_open)
+    @unittest.mock.patch('elitech.device.Path', new_callable=mockpath.MockPath({
         'dev': {
             'hidraw0': b'',
             'hidraw1': b'',
